Add more tests for delete, uncovering bug; #586
This adds more tests for the delete method. One of the areas I'm
concerned about is the interaction between Participant.delete and the
auth system. How does user.ANON behave with a deleted account? In
exploring this I've uncovered a bug in User.from_session_token where the
SQL "session_token = NULL" was translated into the SQLAlchemy expression
"filter_by(session_token=None)", which generates the non-equivalent SQL
expression "session_token is NULL". This is a potential security issue
that needs to be addressed on master immediately. Therefore, I'm
committing on this branch and switching over there.